Reduce volume reduction delay for firefox, add safari compatibility hopefully.
The safari context needs to be webkitAudioContext. We've added
this in, in the hopes that it'll now be usable on IOS.

We've also reduced the time to get to 0 volume to get it
sounding a bit better on Firefox. Firefox had a habit of
letting the notes ring out for too long.